Rename internal/testing to internal/internaltest
