+++ internal/internaltest/alignment.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testing
+package internaltest
 
 /*
 This file contains common utilities and objects to validate memory alignment
