+++ sdk/metric/alignment_test.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"testing"
 
-	ottest "go.opentelemetry.io/otel/internal/testing"
+	ottest "go.opentelemetry.io/otel/internal/internaltest"
 )
 
 // Ensure struct alignment prior to running tests.
